Refactor code
* Storage now manages the loading and saving of tasks from tasklist.txt
* Tasklist.java now contains the tasklist array
